requirements.txt
feedparser==6.0.11
youtube-transcript-api==0.6.2
google-generativeai==0.8.3
requests==2.31.0
python-dotenv==1.0.0
pytz==2024.1